requirements.txt
Flask
google-cloud-datastore
google-auth
requests
google-cloud-storage